pyproject.toml
[tool.poetry]
name = "3scale-deploy"
version = "0.1.0"
description = "Tool to deploy 3scale within running openshift cluster"
authors = ["Miroslav Jaros <mjaros@redhat.com>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{include = "threescale_deploy"}]

[tool.poetry.scripts]
deploy = 'threescale_deploy.installer:cli'

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.3"
boto3 = "^1.26.91"
openshift-client = "^1.0.18"
coloredlogs = "^15.0.1"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pylint = "^2.17.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
